pyproject.toml
[project]
name = "tb-pipeline"
version = "0.1.0"
description = "Phase 1 pipeline: GitHub repos → Harbor dev-env tasks"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.32",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "tenacity>=8.0",
    "docker>=7.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-mock>=3.0", "ruff>=0.4"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pipeline"]

[tool.ruff]
line-length = 100
